requirements.txt
django-environ
djangorestframework
pillow
psycopg2
stripe
boto3
django-taggit
django-allauth
django-rest-auth
django-crispy-forms
crispy-tailwind
django-ckeditor
django-tailwind
markdown
gunicorn
whitenoise
channels
django-redis
sqlparse
asgiref
pytz
mutagen
django-celery-beat
celery
python-dateutil